Moved FIB Pydantic models into 'murfey.util.models', as they will eventually be used by the backend as well
